General class Scraper created
